getEditableNode tests introduced
