+++ UI/src/components/pages/mainPage.jsx
@@ -64,6 +64,7 @@ import { createHistoryUpdater } from "../utils/createHistoryUpdater.js";
 import { undo as undoUtil } from "../utils/undo.js";
 import { redo as redoUtil } from "../utils/redo.js";
 import { handleTabSwitch as handleTabSwitchUtil } from "../utils/handleTabSwitch.js";
+import { getEditableNode} from "../utils/getEditableNode.js";
 
 export const SimulateStateContext = createContext({
   simulateState: "idle",
@@ -135,21 +136,7 @@ export default function Main() {
 
   const ignoreChangesRef = useRef(false);
 
-  const editableNode = useMemo(() => {
-    const selectedNodes = nodes.filter((n) => n.selected);
-    const selectedEdges = edges.filter((e) => e.selected);
-    if (selectedNodes.length === 1 && selectedEdges.length === 0) {
-      const node = selectedNodes[0];
-      if (
-        ["inputNodeSwitch", "inputNodeButton", "outputNodeLed"].includes(
-          node.type,
-        )
-      ) {
-        return node;
-      }
-    }
-    return null;
-  }, [nodes]);
+  const editableNode = useMemo(() => getEditableNode(nodes, edges), [nodes, edges]);
 
   const handleNameChange = (e) => {
     if (!editableNode) return;
